refactor: Add ComprehensionExercise class and update WriteSheets.ipynb
This commit adds a new class called ComprehensionExercise to the ExerciseFactory in Exercise.py. It also updates the WriteSheets.ipynb file to include the necessary code for creating and rendering exercises of type 'Comprehension'. This change allows for the generation of comprehension exercises in the application.

Refactor the WriteSheets.ipynb file to include the necessary code for creating and rendering exercises of type 'Comprehension'. Add the ComprehensionExercise class to the ExerciseFactory in Exercise.py.